requirements.txt
asgiref==3.6.0
Django==4.2
django-admin==2.0.2
django-excel-response2==3.0.5
django-filter==23.1
django-rest-framework==0.1.0
django-six==1.0.5
djangorestframework==3.14.0
excel-base==1.0.4
importlib-metadata==6.6.0
isoweek==1.3.3
Markdown==3.4.3
Pillow==9.5.0
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
screen==1.0.1
six==1.16.0
sqlparse==0.4.4
TimeConvert==3.0.13
tzdata==2023.3
tzlocal==4.3
xlwt==1.3.0
zipp==3.15.0
